Add types folder.
