+++ backend/src/types/auth.types.ts
@@ -1,0 +1,13 @@
+export interface LoginRequest {
+  email: string;
+  password: string;
+}
+
+export interface AuthResponse {
+  success: boolean;
+  error?: string;
+  token?: string;
+  user?: {
+    uid: string;
+  };
+}
