+++ backend/src/types/index.ts
@@ -1,0 +1,5 @@
+// Auth types
+export * from './auth.types';
+
+// Validation types
+export * from './validation.types';
