+++ backend/src/types/validation.types.ts
@@ -1,0 +1,4 @@
+export interface ValidationSchema {
+  email?: boolean;
+  password?: boolean;
+}
